CompletionWords: fix warning messages
The checks were too strict, they were sometimes triggered on text
deletion.

For example:
- Initial buffer contents: " world"
- Insert "hello" at the beginning -> "hello" is added to the
scan_region. New buffer contents: "hello world"
- Quickly delete " world", when "hello" is still in the scan_region.
-> this triggered the warning "'start' iter not adjusted".

Another example:
- Initial buffer contents: "hello "
- Insert "world" at the end.
- Quickly delete "hello ", when "world" is still in the scan_region.
-> this triggered the warning "'text_end' iter not adjusted".

Now it should be fixed.

Ideally unit tests should be written... But the code needs a flush()
function to scan all the text synchronously instead of waiting the
timeout. Note that the examples above assume that an inserted text is
not scanned directly, so a flush() is needed so the unit tests can call
it whenever they want (and so the timeout would be disabled for unit
tests).

https://bugzilla.gnome.org/show_bug.cgi?id=741439