matlab/octave: drop its use for files with mime for the other type.
* both matlab.lang and octave.lang files were currently being used for
files with mimetypes "text/x-matlab" and "text/x-octave".  Octave has
plenty of extra syntax missing from Matlab and thus, the matlab mimetype
should not be used for files identified as octave.  While the opposite
is not true, i.e., Matlab code should run just fine in Octave, in
practice, code that is identified as Matlab is probably aiming at Matlab
compatibility and should avoid using Octave highlight.

* Note that whether a file was written in Octave or Matlab is irrelevant.
If a file was written in Octave but only uses features that are also
present in Matlab, it probably aims at Matlab compatibility and should
be highlighted as Matlab.