Support a current-line-number style in style schemes
Recently we changed the gutter drawing to also draw the current
line background in the gutter. This looks nice for some themes,
like "classic" or gnome-builder's built in theme, but looks bad
in themes like "oblivion" where the line number itself becomes
unreadable. This commit restores the old behavior by default,
but allows color schemes to have

  <style name="current-line-number" background="#yellow"/>

to force a background color for the gutter of the current line
and in particular if the same color of "current-line" is used
then the same effect of the old patch is achieved.

"classic" and "tango" are patched to take advantage of this.