SearchContext: clear search if buffer is destroyed
Ideally we should call clear_search() (and thus removing the idle
function) _when_ the buffer is destroyed, that is, in a GWeakNotify
callback (i.e. use g_object_weak_ref() instead of
g_object_add_weak_pointer() when setting the buffer). But there is a bug
in GObject that prevents doing this, see:
https://bugzilla.gnome.org/show_bug.cgi?id=745678