Don't use gtk_style_context_get_background_color() (deprecated)
The API says to use gtk_render_background() instead, but for the two
cases here it isn't possible.

In GtkSourceCompletion, the background color is applied to a GtkTreeView
cell renderer, which doesn't support CSS.

In GtkSourceStyleScheme, the background color is retrieved to generate
CSS code. AFAIK it isn't possible in CSS to refer to the
background-color value, especially since the background color can be
modified by changing the style scheme, so it isn't always the default
value.