Update the list of CMake builtin variables to that of CMake 3.2.1
Besides the update, the new list also follows the official variable list
(<http://www.cmake.org/cmake/help/v3.2/manual/cmake-variables.7.html>)
in that variable "families" are generalized with a regex. For example,
CMAKE_C_COMPILER and CMAKE_CXX_COMPILER are replaced with CMAKE_\w+_COMPILER.
This might be a bit overzealous, but it's more future-proof (we can't
predict what languages CMake will support in the future, and in fact, the
previous version of the language definition is already missing the variables
for Fortran), and it makes the list easier to update later.

One side effect of the generalization is that some of the listed variables
become redundant. For example, PROJECT_VERSION is subsumed by \w+_VERSION.
I keep them all, though, because that makes the list easier to update.

Some of the variables that were in the list before aren't in the official
builtin variable list and so are removed. I haven't investigated every single
one, but it looks like they're split into the undocumented, the non-builtin,
and the nonexistent (anymore, at least). Either way, I think it's reasonable
to treat the official list as the definitive source for what's builtin.

https://bugzilla.gnome.org/show_bug.cgi?id=746622