style-scheme: use lower priority than APPLICATION for CSS overrides
Using the APPLICATION level priority doesn't allow applications to
override things appropriately. Additionally, it also causes the minimap
to not be able to define a scrubber style without raising the
priority to higher than APPLICATION.

This uses a define so that we can access the value from the minimap
as well.