source-map: make GtkSourceMap inherit from GtkSourceView
The last step in simplifying the minimap was to simply make it inherit
from GtkSourceView. This allows much more customization with regards
to gutters and cleans up the API.

Additionally, we needed to do a few more tricks since we are drawing from
the GtkSourceView's draw vfunc. We no longer invalidate the entire
GtkWidget's GdkWindow. Instead, we only invalidate the old and new
scrubber positions.

This also requires a small change to gtksourcestyle so that we can
ensure our CSS provider is higher than it, but less than the
GTK_STYLE_PROVIDER_PRIORITY_APPLICATION which applications may use.

To simplify the autocleanups, we just use old style declarations like
the other gobjects.