i18n: get rid of intltool
This is the first step. Creating the glade catalog doesn't work anymore,
and strings from XML files are not correctly extracted. This will be
fixed in later commits by using itstool.

intltool is unmaintained, it requires hacks to run 'make distcheck' with
the latest Automake version. gettext has now almost all the features
that we need, except for XML files where itstool can be used.