i18n: use itstool to extract strings from the glade catalog
Glade doesn't support '_name' attributes, it must be 'name'. With the
ITS rules the strings are correctly extracted, but in the xml file it's
less clear what is translated. Maybe a better solution is to run a sed
command to rename _name -> name. Or an even better solution is to
support _name attributes in glade, like we do for .lang files and style
schemes in GtkSourceView.