view: fix and simplify get_leading_trailing()
The trailing iter was not set correctly, it was set one character too
far on the left. The _previous_ character needs to be checked, since the
character pointed to by an iter is the character on the right of the
iter location.

There was other problems in the implementation, like re-using the
"start" variable for... the end of the line. Or hiding the iteration
operation of the loops at the end of the long conditions. The code was
not really clear, and as a proof of that, there was a bug.