view: always do ctrl+backspace behavior for leading spaces
The ctrl+backspace deletes all the leading spaces, if only leading
spaces are on the left of the cursor. This was tight to the
smart-backspace property, but the smart-backspace is more about removing
one indentation level when spaces are used for the indentation.

So it's better to always handle the ctrl+backspace behavior, outside of
the smart-backspace property.

The behavior of ctrl+backspace in GtkTextView is to remove the previous
word. But the previous word can be on a previous line, so it's not a
great default behavior.