pixbuf: handle creating pixbufs for HiDPI resolutions
Under certain HiDPI situations, 1x icons were loaded and rendered with
the scale factor applied. As expected, this resulted in blurry icons.

Of note, were blurry icons in the GtkSourceGutterRendererPixbuf.

This is not the most efficient, but is not terribly different from what
is happening today. We create a new cairo_surface_t for the pixbuf
manually, with the scale applied. gdk_cairo_set_source_pixbuf() also
does this, but without the scale.

We could optimize this in the future by making our pixbuf helper return
cairo_surface_t* rather than GdkPixbuf*.