html: adding embedded-lang-hook context to headings-contexts
Languages embedded im html, whos highlighting is inherited from html.lang
are not hightlighted between the headings-contexts (<title>, <h1>, ...).
This prevents the reader to properly recognize code that is used to
customize the title or headings. For example the Django Template Language
is often used between the <title> tags to customize the title. If the
'{{' and '}}' brackets are not highlighted the reader interprets the
section as text instead of the embedded language. The patch adds the
<context ref="embedded-lang-hook"/> to the headings-contexts to fix this.

https://bugzilla.gnome.org/show_bug.cgi?id=754934