view: do not show undo/redo in context menu if undo/redo is disabled
If the default UndoManager is used and if the max_undo_levels is at 0,
the undo/redo items in the context menu were always insensitive. So it's
better to not show them altogether, instead of taunting the user with
"look there is an undo/redo feature, but you can't use it, haha".