Bracket matching: simplify and rename function cursor_moved()
The name "move_cursor()" was wrong since the function doesn't *move* the
cursor. The cursor has already moved.

And the iter and mark parameters were useless and redundant.

I hope this code was not initially written like that. It's more likely
some leftovers after refactorings. But it means the refactoring was
half-done… great.