Bracket matching: remove another no-longer-needed hack
gtk_source_buffer_set_highlight_matching_brackets() is no longer called
in init(). The value is simply set to TRUE, which is sufficient, since
at that time the buffer is empty, and when some text will be inserted,
the cursor_moved() function will be called.

Anyway, add a unit test to try to set the "highlight-matching-brackets"
property before "tag-table", but it looks like the GTK+ properties are
set before the GtkSourceView ones, so basically it's impossible for
gtk_source_buffer_set_highlight_matching_brackets() to be called before
the tag-table is set. If that changes in the future, the unit test is
there.