Bracket matching: update bracket highlighting on context class changes
In a C file, type:
foo ("(ab)

Before this commit, the ) matched with the first (. It's because the
string context class is not updated directly.

To fix it, the bracket highlighting is now updated when the
highlight-updated signal is emitted.

If it causes performance problems, the bracket highlighting can be
temporarily disabled (e.g. on file loading). Or updating the bracket
highlighting can be done in an idle or timeout function.