GutterRendererText: rely on gtk_render_layout() for the fg color
gtk_render_layout() was already used. Normally it takes care of the
theme, so no need to get the foreground color with
gtk_style_context_get_color(), create manually a PangoAttribute and
apply it to the PangoLayout.

If things break, maybe calling gtk_style_context_set_state() (with
save/restore) is necessary, but it seems to work fine without it.