Style scheme: apply gutter style only to style class
When drawing the gutter, the CSS node is not known. We manually apply
the style class ("left", for example).

There was a bug with the background color in the gutter for the current
line, when the style scheme doesn't define a current-line-number style.
In that case, instead of having the line-numbers background, it had the
same background as the main text area.