CompletionInfo: fix popup window under Wayland
Wayland requires popup windows to be attached to another toplevel
window so that it can be translated either as an xdg-popup or a
subsurface depending if there is a grab associated with the popup or
not.

GtkCompletionInfo uses gtk popup windows type but doesn't set the
transient relationship with the toplevel window, causing the
CompletionInfo window to fail to map under Wayland.

Reuse the attached_to widget to determine the toplevel window and set
the transient relationship prior to map the CompletionInfo window so
that it can work under Wayland as well.

Bugzilla: https://bugzilla.gnome.org/show_bug.cgi?id=759623