css.lang: Add CSS3 keywords and other improvements
The current syntax highlighting file is missing many
css3 properties and values. This patch adds many keywords,
but excludes the more experimental ones. The keywords,
values and at-rules were added in separate contexts
so css1+2 and css3 keywords are separated.
Furthermore the regexes were improved to reduce
mismatches of properties as values and vice-versa.
Some duplicate keywords were removed. The current
lang-file is also missing a definition for floats.
The patch adds the appropriate regex and assigns them
the same color as the 'number with unit' regex.

https://bugzilla.gnome.org/show_bug.cgi?id=663099
Fixes bug #663099