Move ensure highlighting to LAYER_BELOW_TEXT
The new gtk introduced two new drawing layers (BELOW_TEXT and
ABOVE_TEXT) that replace the old BELOW and ABOVE layer. The new
layers work in buffer coordinates and so are easier to use.
This is the first commit in a series to move all our drawing to
the new layers.