Add highlighting for ES6 Template Strings
Template strings are part of ES6 and are currently not
highlighted as strings. This patch adds a context that
gives template strings the same color as single-quoted
and double-quoted strings. The brackets of the template
expressions are given the color of functions. The
expressions themselves are highlighted as functions
too.

https://bugzilla.gnome.org/show_bug.cgi?id=758024
Fixes #758024