styles: avoid using bold for bracket matching
Bold is actually a different font. This might endup being annoying
specially when the normal font size does not match the size of
the bold one and you endup having a lines changing the size when
moving the cursor around the document. This is specially easy to
reproduce on windows.