StyleScheme: add CSS for the left/right/top/bottom classes
Since it's what GtkSourceGutter uses. GtkSourceGutter cannot rely on the
"border" subnode.

The reason why the gutter style was not broken is because it worked by
chance, since we apply the gutter style also to the whole textview node
to fix the color on the corners if the top or bottom window border is
shown.