region: make it public
The implementation could be improved. I added a FIXME comment, and a
GSequence would be a better data structure than a GList. But the region
utility has served GtkSourceView well during 14 years. Instead of
waiting that the code is perfect and waiting another 14 years, it's
better to make the class public now. If someone wants to write a perfect
code and wants to try to get it included in GTK+, at that point we can
deprecate GtkSourceRegion. But I think GtkSourceRegion is now good
enough to be public.