build: dynamic linking for interactive tests
Static linking was used by default, because we linked against the
libtool static convenience libraries. But all tests/*.c files depend
only on public GtkSourceView API, so we can use dynamic linking instead.

The unit tests used to be in tests/, but they are now in testsuite/. In
testsuite/, we need static linking to be able to test private functions.
Now that there are two different directories and two different
Makefile.am's, we can set the right default values for both.