gtksourceview: fix miss-renderings in the grid background
There were situations where we would miss a line or our positioning was
slightly off. This caused some funky updates when the pixelcache decided
to re-render a portion of the visible area.

This uses a more formal "realign" technique that we would use to realign
data-structures to arbitrary offsets (rather than a ^2) which cleans
up the offset calculation a bit too.

The comment was no longer relevant since we have a useful clip region
from the cairo context provided during BELOW_TEXT, so I've removed that
as well.

https://bugzilla.gnome.org/show_bug.cgi?id=765191