gutter: avoid connecting to GtkWidget::draw()
By connecting to the GtkWidget::draw signal, GtkSourceGutter was preventing
GtkWidget to take the fast path by avoiding signal emission. (See
gtkwidget.c for how it checks for pending signal handlers).

This instead adds a private draw function in the gutter that is called by
the view during the render case.

I've verified that this allows test-widget to hit the fast path in
gtkwidget.c.