buffer: add short delay before updating matching-brackets
This adds a low-priority gdk timeout to perform the update to
matching-brackets.

It was discovered, using tests/test-widget, that holding arrow-down or
arrow-up to scroll while matching-brackets is enabled causes stutters in
the drawing. When disabling matching-brackets, this goes away.

By adding a short delay, we reduce the number of updates and invalidations
performed, which keeps us within our deadlines for smooth animations.

The value of 50 was found experimentally, binary searching down from 200
msec (the gtkscrolledwindow animation default) down to 25msec, which gave
unacceptable performance.

https://bugzilla.gnome.org/show_bug.cgi?id=765265