gutter: avoid gtk_text_iter_forward_to_line_end() when possible
When calculating line heights for gutter cells, we don't need precise
access to the beginning of the newline breaks. forward_to_line_end()
will place us just before the \r\n, \n, etc. Instead, we can cheat and
use the line index to jump to the next line and then backwards a char.

This is clearly a micro optimization (about 1-2% on my tests), but as
the line numbers are roughly 2x the cost of the text window, we just
need a few more of these here and there.