map: avoid handling GtkAdjustment signals when hidden
It is often the case that a GtkSourceMap will be packed into the widget
hierarchy, but not visible unless the user has enabled the feature.

This means that we receive a lot of signal notifications when we can't
do anything about it. So instead, toggle the signal activity based on
whether or not we are visible.

https://bugzilla.gnome.org/show_bug.cgi?id=765358