buffer: avoid damage to full buffer when removing matching brackets
When cleaning up matching brackets, gtk_text_buffer_remove_tag() is called
for the entire buffer. This causes a damage in the view causing the entire
pixelcache to be dropped.

Further more, on every ::insert-text we perform the bracket updating
meaning we damage the views on every keystroke.

This walks the tag toggles manually to to reduce the ranges we damage
to just the locations containing the tag.

https://bugzilla.gnome.org/show_bug.cgi?id=765363