FileLoader: move some of the instance variables to a TaskData struct
It's not a good practice to have too many instance variables. So move
the task-related instance variables to a TaskData struct. It's how GTask
is meant to be used. But GtkSourceFileLoader is a bit special, since
there can be at most one task running at a time, otherwise it would
screw up the GtkSourceBuffer content and FileLoader state. So we still
keep the GTask object as an instance variable.

Moving chunk_buffer to the TaskData struct has the nice effect that we
are no longer limited by the maximum size of a GType instance private
data.

Also, make the code more robust. Instead of casting the callback
functions, put the original param types and cast the params inside the
callback function.