tag: use g_object_get() instead of having a private getter
Use g_object_get() like for any GtkTextTag property.

The getter was added recently to improve performances, but if
performance is a problem it's better to create a GtkTextTag (not a
GtkSourceTag) when the GtkSourceTag property is not needed. There is
anyway the 'if (GTK_SOURCE_IS_TAG (tag))' that prevents g_object_get()
from being called for simple GtkTextTags.

Besides, "effects_spaces()" was not a good name. There are other
properties in GtkTextTag that can affect spaces but those were not taken
into account in the getter. And more properties can be added in the
future, which would make effects_spaces() misleading.

So, keep things simple and consistent with GtkTextTag.