view: fix bug in draw_spaces_tag_foreach()
The goal of the function is to find if a GtkSourceTag is present and
that uses the draw-spaces property. It must also return TRUE when
draw-spaces is FALSE, since when it is set, it overrides the global
setting.

There was no bugs in test-space-drawing because it worked by chance,
with the condition at where buffer_has_draw_spaces_tag() is called. But
in the future that code might be modified, or
buffer_has_draw_spaces_tag() could be used at another place, and a bug
would be introduced.

The function name buffer_has_draw_spaces_tag() is normally clear enough
in its meaning...