docs: add links to GRegex and Regular expression syntax page
It doesn't seem possible to link to the glib-regex-syntax page in a
better manner.

This doesn't work:
[Regular expression syntax][glib-regex-syntax]

So a simple link is better than "look at the page titled <foo> in the
GLib reference manual".

And in Devhelp when we click on the link, it is smart enough to stay in
Devhelp and find the right page in the GLib book. So it is not that bad.
I hope the URL will not be modified every 5 years.

Thanks to Stefan Salewski for pointing out the missing info.