CompletionWords: fix comparison between signed and unsigned integers
It's strange that on my machine (Fedora 24) this didn't trigger any
warnings or errors. -Wsign-compare is present on the command line.

g_utf8_strlen() returns a glong. minimum_word_size is a guint. So cast
the latter to glong. This should make the error to disappear, but there
are maybe other places in GtkSourceView where a glong is compared to a
guint. So finding a way to reproduce the warning/error would be better.

https://bugzilla.gnome.org/show_bug.cgi?id=767254