view: connect to ::highlight-updated signal to update highlight
There was a good reason behind
commit 949cd128064201935f90d999544e6a19f8e3baa6: the code was not clear.
It was not evident to know why that code was useful.

If the title of this commit doesn't make any sense, it's because the
::highlight-updated signal has a bad name. When the ContextEngine sends
it, the _highlight_ is actually not yet updated.

Now with the new implementation and the comment in the code, I hope
it'll be a bit clearer. But it would be nice to do something about the
signal name.

https://bugzilla.gnome.org/show_bug.cgi?id=767565