readme: simplify install procedure
The README is inside the tarball, so by the time someone reads the
README, the tarball is already extracted. Besides, if someone wants to
compile GtkSourceView from source, we can expect that extracting the
tarball is something that he/she can do.

./configure && make && make install is the standard procedure, no need
to detail everything.

And replace % by $ as the shell prompt. Bash is usually the shell
installed by default.