Always build words completion provider
The words completion provider is not a lot of code and doesn't pull up
additional dependencies.

Currently, applications need the following #include to use the words
completion provider:
 #include <gtksourceview/completion-providers/words/gtksourcecompletionwords.h>

By always building the completion provider, it'll be easier to #include
that header in gtksource.h (it'll be done in a later commit), so that
applications have just one header to #include, it'll be simpler.