Include words completion provider header in gtksoure.h
Now that the completion provider is always built (see previous commit),
we can easily #include the header in the main header of GtkSourceView,
gtksource.h. So that applications have just one header to #include, it's
simpler.

For generating GtkSource-3.0.gir, gtksource.h must be the first header
in the list, otherwise there are warnings.