latex.lang: remove "name" style
For example in:
\documentclass[a4paper,11pt]{report}

"report" was highlighted with the "name" style.

This doesn't add any value. The goal of syntax highlighting is not to
have the most colorful text as possible. Highlighting the command name
is sufficient (in the example, \documentclass).

Besides, the value inside the curly braces was highlighted, but there
was no checks to see if the value is correct. Usually, when something is
highlighted, it means that it is correct (except with a red background
or something).