region: rename functions to add/subtract/intersect a subregion
Now that add/subtract/intersect_region() functions exist, it was not
symmetric. So change the function names to be explicit.

Another possible name is "range" instead of "subregion", but "subregion"
is already used by gtk_source_region_iter_get_subregion(), so using the
same terminology is better.

The GtkSourceRegion has been made public during this development cycle,
so it is still allowed to change the API. Once the 3.22 version is
released, API stability will be guaranteed.