view: improve docs of get_space_drawer()
In the future, there will maybe be a set_space_drawer() function, for
example to override a vfunc of SpaceDrawer and set that. Code written
before adding the set function will still continue to work fine, so it
will not be an API break to add the set function and remove the
guarantee that the returned SpaceDrawer is the same.