SpaceDrawer: cairo_stroke() each character
draw_nbsp_at_pos() already called cairo_fill() or cairo_stroke(). In
case of cairo_fill() (for a narrowed nbsp), it affected the previous
char. So now, cairo_stroke() is called for each character, so that
drawing a character doesn't affect the drawing of another.

There is a screenshot of the bug at:
https://bugzilla.gnome.org/show_bug.cgi?id=773214