Bump API version to 4
The previous API version was "3.0". Now instead of being "4.0", it is
just "4" because it is more logical: from the package version (of the
form major.minor.micro), only the major version gathers
backward-compatibility information, i.e. the API version.

This is what the GNOME Programming Guidelines recommend:
https://developer.gnome.org/programming-guidelines/unstable/parallel-installability.html.en

It also permits to use the GSV_API_VERSION variable in the
gtksourceview/ directory.