Visual Studio builds: Prepare for 4.x
Rename the projects and update things as necessary so that these projects
can be used in parallel with the 3.x projects, and make sure the autotools
items for generating the complete projects and scripts are made up-to-date
as a result.  This attempts to not hardcode version info in the autotools
items as well.

https://bugzilla.gnome.org/show_bug.cgi?id=774093