Visual Studio builds: Add support for Visual Studio 2017
This adds support for Visual Studio 2017 by updating the autotools
scripts to copy the 2010 projects and updating items as necessary to
obtain the 2017 projects.

Note that the toolset version string format changed for Visual Studio
2017, so a custom toolset version string is allowed and used if
specified, otherwise it will be generated as it were before.

Note also that Visual Studio 2017 aims to be compatible with Visual Studio
2015 on the CRT level, so 2017-built binaries should work with 2015-built
binaries without problems.