rust.lang: fix invalid-char context matching rust's lifetime syntax
Lifetimes in rust use a single quote as part of their syntax.
When 2 or more lifetimes appear on a line, part of the line is
(incorrectly) highlighted by the invalid-char context.

Example:

    struct SomeObject<'a, T: 'a> {}

Before this patch, everything between the quotes on this line is
highlighted as invalid-char.

This patch prevents the invalid-char context in rust.lang from
matching single quotes which are part of the lifetime syntax,
by abandoning a match for invalid-char if the closing quote is
immediately followed by an identifier.

https://bugzilla.gnome.org/show_bug.cgi?id=779236