Visual Studio builds: Move projects to win32/
This moves the projects from build/win32/vsX to win32/vsX so that one will
need to go down one less level into the source tree to reach the project
files, which was suggested some time ago.  This will also enable the
common NMake Makefile and autotools modules in win32/ to be in sync with
the latest ones in GLib and G-I master.

This will also clean up the Visual Studio 201x projects as we can reduce
the number of configs for custom build steps.