Visual Studio builds: Add support for Visual Studio 2017
This adds support for Visual Studio 2017 projects by using the new
win32/Makefile-newvs.am that was updated from the last commit, so that
the 2010 projects can be copied and updated accordingly for the 2017
projects.

Note that the format of the toolset version string changed for Visual
Studio 2017, so allow and use a custom toolset version string when
specified, otherwise generate it as it was before.

Note also that Visual Studio 2017 aims to be compatible with Visual Studio
2015 on the CRT level, so one should be able to use 2017-compiled binaries
with 2015-compiled binaries without problems.