move-lines: deprecate copy parameter
The name of the signal is *move*-lines. Not copy. Anyway, how it is
implemented it doesn't make sense, that's a very strange operation to do
with a keyboard shortcut. See the removed unit test. It's better to do
only one thing but do it well. copy=true was anyway buggy when several
lines were selected (but it should now be fixed with the previous
commit).

Print also a warning if 'count' is different than 1 or -1.