build: apply correctly HIDDEN_VISIBILITY_CFLAGS
It was applied only to the real libtool library. But all the source code
is compiled into the helper libtool libraries libgtksourceview-core.la
and libgtksourcecompletionwords.la. So the -fvisibility=hidden was
actually not applied!

The unit tests are still able to call private symbols because the
-export-dynamic LDFLAG is applied only to the real library.

When running:
$ nm -D libgtksourceview-4.so | grep T
before and after this commit, the only difference is that
gtksourceview_get_resource() is no longer exported, which is fine
because it was exported by mistake, the function permits to access the
internal GResources.

Thanksfully, the other private symbols were marked with G_GNUC_INTERNAL,
otherwise all private symbols would have been exported.