language-specs: remove translatable strings from new *.lang files
Keep only the _section translatable, those make sense to translate and
they are already translated.

The language names (Swift, Kotlin and Maxima) should not be translated,
for example if Maxima is translated the user will certainly be confused…

The style names are not very useful to translate, because the user never
sees those strings! It is used for example by a gedit plugin to
create/edit style schemes. The style names are probably hard to
translate anyway, and I'm sure that it is often translated in a way that
the user doesn't understand what it means… So, since those strings are
extremely rarely used, and since they are hard to translate without
context, it is better to not translate them at all. It's like GObject
properties.

The *.lang file format doesn't support well translatable strings.
Ideally each translatable string should have its own tag, with the
translatable="yes" property, the possibility to add a context, etc.

The Kotlin, Maxima and Swift *.lang files will be backported on
GtkSourceView 3.24, where we must not add new translatable strings. So
this commit eases the process and will permit to cherry-pick new commits
for those *.lang files without conflict.

It is planned to remove translatable strings from all *.lang files in
the future (except the _section), so this is just a first step for the
new *.lang files that will be backported.