language-specs: do not mark strings for translation
Keep only the _section translatable, those make sense to translate.

The language names (e.g. Swift, Kotlin, Maxima) should not be
translated, for example if Maxima is translated the user will certainly
be confused…

The style names are not very useful to translate, because the user never
sees those strings! It is used for example by a gedit plugin to
create/edit style schemes. The style names are probably hard to
translate anyway, and I'm sure that it was often translated in a way
that the user didn't understand what it means… So, since those strings
are extremely rarely used, and since they are hard to translate without
context, it is better to not translate them at all. It's like GObject
properties.

Also, the *.lang file format doesn't support well translatable strings.
Ideally each translatable string should have its own tag, with the
translatable="yes" property, the possibility to add a context, etc.