MSVC builds: Adapt to Pythonfied glib-mkenums
glib-mkenums has been re-written in Python in the latest GLib that will
be in the next stable release 2.54.0, so we need to adapt to it.

The issue is that the standard Windows cmd.exe (where the Visual Studio
projects make use of) does not support shebang lines, and we still
support GLib versions that use the PERL glib-mkenums, we need to first
try to run that script with Python (which we already use during the
build) and see whether that was successful by seeing whether the
generated file is not empty (as calling the PERL glib-mkenums would
cause an error and results in the generated file being 0 in size).

If calling glib-mkenums with Python failed, call that script with PERL,
as we did before.