build: Integrate introspection builds in MSVC projects
Allow introspection builds to be carried out in the Visual Studio
projects directly, by using a project which wraps calls to the NMake
Makefiles that is responsible for building the introspection files.

Note that as in the GTK+ stack, this is not built (nor cleaned or
rebuilt) by default, so the "gtksourceview-introspect" project must be
explicitly selected and built (or cleaned or rebuilt) for the
introspection files to be built and installed to appropriate locations
under $(GlibEtcInstallRoot) from within the projects.  This will do a
build of the GtkSourceview DLL and the "install" as usual before
generating the introspection files.

In order for this to work, one needs to:

-Have GObject-Introspection built and "installed" under
 $(GlibEtcInstallRoot).  Check that PythonDir (or PythonDirX64 for x64
 builds) points to the same Python installation that was used to build
 GObject-Introspection.
-Have the introspection files for GTK+-3.x under appropriate locations
 under $(GlibEtcInstallRoot)