view: indent: skip only *completely* empty lines
In commit e9f3d760dc6bbb59aae9f8b19716deeeed96e338 (the commit that
added the indent function in 2006) we can easily see that only
*completely* empty lines are skipped, so this was the original intent.

And when thinking about it, it makes sense, as the new comment explains
in the code.

This bug was introduced by
commit 874a07f0202722607146db39b32f907a2de13b2e in 2007. It doesn't make
sense to skip lines containing only tabs but not lines containing only
spaces (or tabs + spaces). 10 years later the bug is reported (and
fixed).

https://bugzilla.gnome.org/show_bug.cgi?id=787268