Include: gtksourceview-i18n.h -> glib/gi18n-lib.h where sufficient
Or simply remove the #include where gettext is not used.

See the previous commit that removes the ENABLE_NLS option. Now
including glib/gi18n-lib.h is sufficient in most places, which is also
more standard for a GNOME library, so it is easier for contributors.

Also fix #include order where it was wrong, see:
https://wiki.gnome.org/Projects/GTK%2B/BestPractices/GlibIncludes