License headers: remove filenames
The LGPL doesn't require it. When renaming a file, we often forgot to
update the filename. And when the namespace will be renamed, it'll be
one less thing to care about.

Done with this command (or a few other variants):
$ sed -i 's/^ \* gtksource.*\.[ch]$/ */' *.[ch]
and then by looking at the diff to see if it didn't remove something
else.