cmake.lang: add highlighting for environment variable references
The syntax is nearly the same as for regular variable references, and
I would've just merged the contexts, but builtin variable names have
no special meaning in environment variable references, so I opted to
use a separate context to implement that distinction.

There are actually special environment variable names that CMake
recognizes, but I haven't implemented support for them.

https://bugzilla.gnome.org/show_bug.cgi?id=790345